[www] Try harder to get exact data out of rrdtool fetch
rrdtool always give you one extra row of data beyond what you query for (v1.4.5 currently).  If now is the exact time a new chunk was added, we'll also get a second row full of NaN because data hasn't been added yet.  We play it safe and try to eat those NaN rows, without backing up the window so far that it it extends beyond the beginning of the database, which will bump up the rate to the next largest value.