[hm] Add GrillPid servo mode output, all modes use TIMER1 now
Because we normally use digital pin 3 for fan output, there was an
interaction between the PWM output and the tone() output, both of which
use the same timer but in different modes.

This new code always uses TIMER1 and controls the PWM in software, so the
blower output can be in any pin.

This also adds the 50Hz servo output mode, using fanspeed min/max to set
servo pulse duration in 10s of microseconds (e.g. 150 = 1500usec)