[lm] Add an autobackup of the running RRD file that's restored on boot
There are two autobackup intervals:
Pit < 150 or no probe = "inactive"
Pit >= 150 = "active"

The interval should not be set so low that it wears out the SD card
with constant overwrites.

Before a new RRD is created, the autobackup is checked and if the last
update was in the past, then that file is copied into place and continued
rather than creating a new file.