[lm] Store first boot configuration in sd partition default-config.tar.gz
This isn't actually a 100% default config as some services have already
started and created some additional files. However, we can't run any earlier
because then we wouldn't have all the files needed to wipe the system when
restoring the backup.

For example, if we do the default-config backup before rc.firewall is ever
created then when we restore the backup it won't overwrite the possibly-
broken version of the default-config the wipe is trying to fix.