[hm] Only reduce, not fully cut, the I term when temperature is reached.
I'm seeing a lot of graphs where reaching the temperature cuts the PID
output so much that it can't recover before Lid Detect kicks in, turns
off the output and the cycle repeats. This change is a band-aid to that.
Failing not knowing how much I is needed, I'll take 1/4 as a starting
point. Experimentation may determing 1/2 or 3/4 might be even better.