[hm] Fix fan speed changing while alarm is sounding
tone() uses TIMER2 to generate its output and modifies the waveform
generation to do so. Because the blower uses hardware PWM on TIMER2,
this makes the blower run full on or off while the alarm is ringing.

This code fixes the alarm output at 4KHz and uses TIMER1's COMPA
vector to do the timing (same timer used for the servo output).
The new code modifiest the timer mode from normal to CTC with ICR1
as TOP which reduces the amout of code in the COMPB vector to a
single line.

Thanks to Peter F for finding the bug