[hm] FanByServo ( PIDFLAG 5)
Use PidOutputFlag 5 to enable.
Logic: I'm considering the fan an air booster with 10 gear speeds. When
servo gets over 70 or under 30 the fan speed is shifted. Can only shift
once every 30 seconds to give PID a chance to react. Also the integrator
gets moved 5% to compensate for the airflow change