[hm] Prevent low current fans from turning off in voltage mode at the bottom of the PWM range
-- Thanks to R-Burk for discovering and finding workable solutions

When attempting to do voltage mode fan control, if the current draw is too low
the output duty cycle will throttle down to lower the voltage. If throttled
back too far (0), the PWM turns off completely which causes the voltage
to drop hard and the PWM to shoot back up to compensate.

The simple solution is to just not turn off at all when there's any desired output.