[hm] Change servo behavior to be fixed speed and stay energized 2s after arriving
The new mechanism moves the servo at a constant speed of 15us per every
pulse duration and it takes as long as it takes to get there. Once
the servo arrives at its destination, it will always stay energized there
for 2 full seconds (100 periods) before shutting down, regardless of
distance traveled. If there is a change in destination the servo will slew
toward that target rather than the initial target and the travel time
changes accordingly.