[hm] Reduce servo jitter from TIMER1_CAPT interrupt latency and calculation time
ISR latency is pretty bad so it causes the servo to be jumpy. This code
shifts the end of the pulse by the current value of the timer at the
pulse start which means only the "turn off" interrupt is affected now.

Thanks to Rob Burk for this clever fix!