[hm] Greatly increase response of button input and disable input for 250ms after each event
The ADC loop in grillpid samples 255 + 3 or 48 discard samples for each
6 ADC inputs. Each sample is 0.105ms so 27.09ms per ADC or 162.54ms per
loop. This high latency between button means a press can be missed,
especially if it was being pressed during the 27ms poll of the button
input (which would be discarded as a transition) so the button would
have to be held at least 320ms to be detected.

The new code polls the buttons for just 4 samples in between each of the
other full-oversampling ADC readings. This creates a 27ms max latency
between button polls and because the sampling period is only 0.42ms
there is 50x less of a chance of catching an edge (which is on the order
of 100ns). Reducing this to 1 or 2 samples may be beneficial if that is
found to be a problem.

As a side effect of the high button latency, the user would hold down
the button for longer to get it to respond, which also would create key
repeats because they would need to hold it for 160-320ms but then
release it in under 250ms. The new code enforces a 250ms lockout of all
activity after any button event, altough the responsiveness of the
buttons now mitigate this issue.