[lm] Add the ability to save PNG images of live graphs as well as download CSV as attachment
Adds a /lm/dluri controller which takes a data URI as a parameter and spits it back as
content-disposition attachment which prompts the user to save as. Also removes the PNG toggle
on the archive page in leiu of using this new portable download method.

This changes the graph to always use a canvas, not sure if that will have any side effects.