[hm] Buttons now checked 250ms after last detection instead of waiting after every state change
Previously there would be a 250ms delay after every state change before
we'd consider a button again, which is fine if pressing a button is what
caused the state change.

However on the home screen, there are timeouts occuring every second
which counts as a state change. This locks out the user from pressing
any button 1/4 of the time. This change now always looks at the button
state regardless of when the last state transition occurred and only
locks out changes for 250ms after a button detection (!= BUTTON_NONE).

It has the side effect of delaying a timeout if a button is pressed but
I don't think that will be a problem because the button should trigger
its own state change or update the display which would mask the 250ms
lag in timeout.