[hm] Add menu button longpress support / PID disabled function
Menus can now register for longpress actions by including which buttons
support longpress in their state in the menu_definition_t::longpress
field. Buttons which do not have longpress handlers in the current state
execute on button down and can have repeat counts. If the button is
pressed for less than the longpress duration, the regular button handler
fires, else (button | BUTTON_LONG) fires. Can be on auto transitions or
in state hander_t functions.

This also includes a hander in the home menu for turning on / off PID
output completely. Hey, I needed to test it! And fixes a couple minor
bugs:
-- Allow setpoints of just units without switching the device to manual
mode
-- Fixes LCD not updating for 1 second when leaving menus when no food
probes are plugged in
-- Fixes button repeat count persisting between state changes (don't
think this was an issue though)