[wrt] Fix for reading entirety of STDIN while trying to stream file through CGI boundary
The uhttpd embedded LUA processor would read the entirety of an uploaded file when a single uhttpd.read()
call was made. The code would read 1KB at a time and concatenate it over and over until it read the whole
file (up to 29MB in our case). This would consume massive amounts of RAM and take a very long time
causing Bad Gateway messages when the server timed out.

The fix now only reads the requested amount so the file is streamed in 4KB chunks. Trying to get this
integrated upstream, the patch location in install.sh is pretty hacky.

Also added the 900 dnsmasq patch to the install for now.