[3d] Adjust LCD down ~0.5mm
The LCD is actually not centered on its PCB which I failed to recognize
when doing the design. This adjusts its window properly and also sets
its anchor based on the actual board layout. Overall this moves the
entire LCD/Button/LED section closer to the front by about 0.5mm and
seems to fit a little better.