[lm/www] Add support for fanpct and servopct from HMSU updates
Database will have to be restarted to store the new values,
but linkmeterd and the webui should be tolerant of both running
an old AVR firmware with fewer fields, and displaying databases
that do not have the f and s fields.

The home page also now shows Fan and Blower speed as text in the
Output bar, but only if they are different from the Output percent.